商店可以買東西了、可以裝備東西、把主角統一

+++ final project/Assets/scripts/tower_thing/Mapmanager.cs
@@ -13,7 +13,7 @@ public class Mapmanager : MonoBehaviour
         }
         else
         {
-            loader.LoadMaps(Random.Range(0,4));
+            loader.LoadMaps((int)Random.Range(0,4));
         }
         
     }
